Remove all sourceMappingURL comment types
The removeComments function was only removing the sourceMappingURL for embedded (inlined) sources and not map file URLs. This fix removes both comment types.

Because Browserify through browser-pack depends on this module, this ensures that only a single sourceMappingURL is present in the final bundle file.

When multiple maps were present, gulp-sourcemaps, Safari, and other developer tools would only read the first map and miss the combined one at the end.