Integrate path-is-absolute directly and fix dependencies
The previous version took a dependency on path-is-absolute which requires node engine >=0.10.0. The component doesn't seem to actually require it, so in order to avoid moving combine-source-map to a newer engine the component is included directly as a lib.

Additionally, other component versions were updated but used the the ^ syntax instead of ~ which is safer so this has been corrected.